# api/openapi.yaml
openapi: 3.1.0

info:
  title: Questionnaire API
  version: 0.1.0
  description: API for managing quizzes, questions and challenges.

servers:
  - url: /api/v1

paths:
  /quizzes:
    post:
      summary: Create quiz
      operationId: createQuiz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuiz"
      responses:
        "201":
          description: Quiz created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"

  /quizzes/{quizId}:
    parameters:
      - $ref: "#/components/parameters/QuizId"

    get:
      summary: Get quiz
      operationId: getQuiz
      responses:
        "200":
          description: Quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Update quiz
      operationId: updateQuiz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuiz"
      responses:
        "204":
          description: Quiz updated
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Delete quiz
      operationId: deleteQuiz
      responses:
        "204":
          description: Quiz deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /quizzes/{quizId}/questions:
    parameters:
      - $ref: "#/components/parameters/QuizId"

    post:
      summary: Add question to quiz
      operationId: addQuestionToQuiz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddQuestionToQuiz"
      responses:
        "204":
          description: Question added to quiz
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /challenges:
    post:
      summary: Create challenge
      operationId: createChallenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChallenge"
      responses:
        "201":
          description: Challenge created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Challenge"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /challenges/{challengeId}:
    parameters:
      - $ref: "#/components/parameters/ChallengeId"

    get:
      summary: Get challenge overview
      operationId: getChallenge
      responses:
        "200":
          description: Challenge overview with question list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeOverview"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Delete challenge
      operationId: deleteChallenge
      responses:
        "204":
          description: Challenge deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /challenges/{challengeId}/question/{questionId}:
    parameters:
      - $ref: "#/components/parameters/ChallengeId"
      - $ref: "#/components/parameters/QuestionId"

    get:
      summary: Get question from challenge
      operationId: getChallengeQuestion
      responses:
        "200":
          description: Challenge question with answers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeQuestion"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /challenges/{challengeId}/close:
    parameters:
      - $ref: "#/components/parameters/ChallengeId"

    post:
      summary: Close challenge
      operationId: closeChallenge
      responses:
        "204":
          description: Challenge closed
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /challenges/{challengeId}/answers:
    parameters:
      - $ref: "#/components/parameters/ChallengeId"

    post:
      summary: Submit an answer
      operationId: submitAnswer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAnswer"
      responses:
        "204":
          description: Answer recorded
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /challenges/{challengeId}/results:
    parameters:
      - $ref: "#/components/parameters/ChallengeId"

    get:
      summary: Get challenge results
      operationId: getChallengeResults
      responses:
        "200":
          description: Results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Result"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /questions:
    post:
      summary: Create question
      operationId: createQuestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestion"
      responses:
        "201":
          description: Question created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"

  /questions/{questionId}:
    parameters:
      - $ref: "#/components/parameters/QuestionId"

    get:
      summary: Get question
      operationId: getQuestion
      responses:
        "200":
          description: Question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Update question
      operationId: updateQuestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuestion"
      responses:
        "204":
          description: Question updated
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Delete question
      operationId: deleteQuestion
      responses:
        "204":
          description: Question deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  parameters:
    QuizId:
      name: quizId
      in: path
      required: true
      schema:
        type: string
        format: uuid

    ChallengeId:
      name: challengeId
      in: path
      required: true
      schema:
        type: string
        format: uuid

    QuestionId:
      name: questionId
      in: path
      required: true
      schema:
        type: string
        format: uuid

  responses:
    NotFound:
      description: Resource not found

    Forbidden:
      description: Not allowed to access the resource

  schemas:
    Visibility:
      type: string
      enum:
        - public
        - private

    QuizQuestion:
      type: object
      required:
        - questionId
        - position
      properties:
        questionId:
          type: string
          format: uuid
        position:
          type: integer
          minimum: 0

    Quiz:
      type: object
      required:
        - id
        - owner
        - title
        - description
        - visibility
        - questions
      properties:
        id:
          type: string
          format: uuid
        owner:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        visibility:
          $ref: "#/components/schemas/Visibility"
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"

    CreateQuiz:
      type: object
      required:
        - title
        - description
        - visibility
      properties:
        title:
          type: string
        description:
          type: string
          default: ""
        visibility:
          $ref: "#/components/schemas/Visibility"
        questions:
          type: array
          default: []
          items:
            $ref: "#/components/schemas/QuizQuestion"

    UpdateQuiz:
      type: object
      required:
        - title
        - description
        - visibility
        - questions
      properties:
        title:
          type: string
        description:
          type: string
        visibility:
          $ref: "#/components/schemas/Visibility"
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"

    Access:
      type: string
      enum:
        - public
        - invited

    ChallengeState:
      type: string
      enum:
        - scheduled
        - running
        - finished

    ChallengeAnswer:
      type: object
      required:
        - id
        - description
        - position
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
        position:
          type: integer
          minimum: 0

    ChallengeQuestionSummary:
      type: object
      required:
        - id
        - description
        - position
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
        position:
          type: integer
          minimum: 0

    ChallengeQuestion:
      type: object
      required:
        - id
        - description
        - position
        - answers
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
        position:
          type: integer
          minimum: 0
        answers:
          type: array
          items:
            $ref: "#/components/schemas/ChallengeAnswer"

    ChallengeQuiz:
      type: object
      required:
        - quizId
        - title
        - questions
      properties:
        quizId:
          type: string
          format: uuid
        title:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/ChallengeQuestion"

    ChallengeOverview:
      type: object
      required:
        - id
        - quizId
        - quizTitle
        - startsAt
        - access
        - invitees
        - state
        - questions
      properties:
        id:
          type: string
          format: uuid
        quizId:
          type: string
          format: uuid
        quizTitle:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
        access:
          $ref: "#/components/schemas/Access"
        invitees:
          type: array
          items:
            type: string
            format: uuid
        state:
          $ref: "#/components/schemas/ChallengeState"
        questions:
          type: array
          items:
            $ref: "#/components/schemas/ChallengeQuestionSummary"

    Challenge:
      type: object
      required:
        - id
        - quiz
        - startsAt
        - access
        - invitees
        - state
      properties:
        id:
          type: string
          format: uuid
        quiz:
          $ref: "#/components/schemas/ChallengeQuiz"
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
        access:
          $ref: "#/components/schemas/Access"
        invitees:
          type: array
          items:
            type: string
            format: uuid
        state:
          $ref: "#/components/schemas/ChallengeState"

    CreateChallenge:
      type: object
      required:
        - quizId
        - startsAt
        - access
      properties:
        quizId:
          type: string
          format: uuid
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        access:
          $ref: "#/components/schemas/Access"
        invitees:
          type: array
          default: []
          items:
            type: string
            format: uuid

    SubmitAnswer:
      type: object
      required:
        - questionId
        - answerId
      properties:
        questionId:
          type: string
          format: uuid
        answerId:
          type: string
          format: uuid

    Result:
      type: object
      required:
        - userId
        - points
      properties:
        userId:
          type: string
          format: uuid
        points:
          type: integer
          minimum: 0

    Answer:
      type: object
      required:
        - id
        - description
        - position
        - correct
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
        position:
          type: integer
          minimum: 0
        correct:
          type: boolean

    Question:
      type: object
      required:
        - id
        - owner
        - description
        - visibility
        - labels
        - answers
      properties:
        id:
          type: string
          format: uuid
        owner:
          type: string
          format: uuid
        description:
          type: string
        visibility:
          $ref: "#/components/schemas/Visibility"
        labels:
          type: array
          items:
            type: string
        answers:
          type: array
          minItems: 0
          items:
            $ref: "#/components/schemas/Answer"

    CreateQuestion:
      type: object
      required:
        - description
        - visibility
        - answers
      properties:
        description:
          type: string
        visibility:
          $ref: "#/components/schemas/Visibility"
        labels:
          type: array
          default: []
          items:
            type: string
        answers:
          type: array
          minItems: 0
          items:
            $ref: "#/components/schemas/Answer"

    UpdateQuestion:
      type: object
      required:
        - description
        - visibility
        - answers
      properties:
        description:
          type: string
        visibility:
          $ref: "#/components/schemas/Visibility"
        labels:
          type: array
          default: []
          items:
            type: string
        answers:
          type: array
          minItems: 0
          items:
            $ref: "#/components/schemas/Answer"

    AddQuestionToQuiz:
      type: object
      required:
        - questionId
        - position
      properties:
        questionId:
          type: string
          format: uuid
        position:
          type: integer
          minimum: 0